Move plotting to utils
